requirements.txt
asgiref==3.8.1
Django==5.1.4
django-axes==7.0.1
django-ckeditor==6.7.2
django-ckeditor-5==0.2.15
django-crispy-forms==1.10.0
django-environ==0.11.2
django-formtools==2.5.1
django-js-asset==3.0.1
django-otp==1.5.4
django-phonenumber-field==8.0.0
django-tinymce==4.1.0
django-two-factor-auth==1.17.0
pillow==11.1.0
pypng==0.20220715.0
qrcode==7.4.2
sqlparse==0.5.3
style==1.1.0
typing_extensions==4.12.2
update==0.0.1
